Include EFF wordlist in repo
Following a mail exchange it turns out that the EFF wordlist is CC BY 3.0 US,
see https://www.eff.org/copyright, so we can perfectly embed it in the repo.

Hence the build script is gone now, and we can build without network access
provided that all dependencies are present.

Yay <3